move declarations
