+++ src/models/TsOldDeclaration.ts
@@ -21,10 +21,6 @@ export class TsClassDeclaration extends TsExportableDeclaration {
     }
 }
 
-export class TsMethodDeclaration extends TsDeclaration {
-    
-}
-
 export class TsFunctionDeclaration extends TsExportableDeclaration {
     public getItemKind(): CompletionItemKind {
         return CompletionItemKind.Function;
