+++ src/models/TsResolveInformation.ts
@@ -1,6 +1,6 @@
 import {TsImport} from './TsImport';
 import {TsExport} from './TsExport';
-import {TsDeclaration} from './TsDeclaration';
+import {TsDeclaration} from './TsOldDeclaration';
 
 export interface TsResolveInformation {
     imports: TsImport[];
